\name{AddVariance}
\alias{AddVariance}
\title{
Adds variance
}
\description{
This function adds variance if necessary
}
\usage{
AddVariance(x, k, parameters)
}
\arguments{
  \item{x}{
The supplied data matrix. 
}
  \item{k}{
The number of variables.
}
  \item{parameters}{
The data and program parameters
}
}
\details{
Called by higher-order functions; users do not need to call this 
function directly.
}
\value{
Data with necessary variance added
}
\author{
John Ruscio <ruscio@tcnj.edu> and Shirley Wang <shirleywang@g.harvard.edu>
Maintainer: John Ruscio <ruscio@tcnj.edu>
}
