\name{AssignMAMBAC}
\alias{AssignMAMBAC}
\title{
Assigns variables to input/output for MAMBAC procedure
}
\description{
This function assigns variables to input/output configurations for 
MAMBAC analysis.
}
\usage{
AssignMAMBAC(parameters)
}
\arguments{
  \item{parameters}{
The data and program parameters
}
}
\details{
Called by higher-order functions; users do not need to call this 
function directly.
}
\value{
Input/output variables per curve
}
\author{
John Ruscio <ruscio@tcnj.edu> and Shirley Wang <wangs7@tcnj.edu>
Maintainer: John Ruscio <ruscio@tcnj.edu>
}