\name{CalculateBaseRates}
\alias{CalculateBaseRates}
\title{
Calculates and returns base rate estimates for taxometric analysis
}
\description{
This function calculates and reports base rate estimates for taxometric
analysis.
}
\usage{
CalculateBaseRates(x.results, parameters)
}
\arguments{
\item{x.results}{
Empirical data results
}
  \item{parameters}{
The data and program parameters
}
}
\details{
Called by higher-order functions; users do not need to call this 
function directly.
}
\value{
This program returns nothing, and provides text output only.
}
\author{
John Ruscio <ruscio@tcnj.edu> and Shirley Wang <shirleywang@g.harvard.edu>
Maintainer: John Ruscio <ruscio@tcnj.edu>
}
