\name{CalculateFitDensities}
\alias{CalculateFitDensities}
\title{
Calculates fit
}
\description{
This function calculates fit for L-Mode curves
}
\usage{
CalculateFitDensities(shift, data)
}
\arguments{
  \item{shift}{
Horizontal shift
}
  \item{data}{
Curves for empirical and comparison data
}
}
\details{
Called by higher-order functions; users do not need to call this 
function directly.
}
\value{
Fit value
}
\author{
John Ruscio <ruscio@tcnj.edu> and Shirley Wang <shirleywang@g.harvard.edu>
Maintainer: John Ruscio <ruscio@tcnj.edu>
}
