\name{CalculateKurtosis}
\alias{CalculateKurtosis}
\title{
Calculates kurtosis
}
\description{
This function calculates the sample kurtosis of a distribution
}
\usage{
CalculateKurtosis(x)
}
\arguments{
  \item{x}{
The data vector
}
}
\details{
Called by higher-order functions; users do not need to call this 
function directly.
}
\value{
The sample kurtosis of x
}
\author{
John Ruscio <ruscio@tcnj.edu> and Shirley Wang <shirleywang@g.harvard.edu>
Maintainer: John Ruscio <ruscio@tcnj.edu>
}
