\name{CalculateMAMBAC}
\alias{CalculateMAMBAC}
\title{
Calculates MAMBAC curve
}
\description{
This function calculates one MAMBAC curve
}
\usage{
CalculateMAMBAC(input, output, parameters)
}
\arguments{
  \item{input}{
Input indicator
}
  \item{output}{
Output indicator
}
  \item{parameters}{
The data and program parameters
}
}
\details{
Called by higher-order functions; users do not need to call this 
function directly.
}
\value{
One MAMBAC curve
}
\author{
John Ruscio <ruscio@tcnj.edu> and Shirley Wang <shirleywang@g.harvard.edu>
Maintainer: John Ruscio <ruscio@tcnj.edu>
}
