\name{CalculateMAXSLOPE}
\alias{CalculateMAXSLOPE}
\title{
Calculates MAXSLOPE curve
}
\description{
This function calculates one MAXSLOPE curve
}
\usage{
CalculateMAXSLOPE(x, curve)
}
\arguments{
  \item{x}{
The data matrix
}
  \item{curve}{
Curve number
}
}
\details{
Called by higher-order functions; users do not need to call this 
function directly.
}
\value{
List object with one MAXSLOPE curve:
\item{curve.x}{x values}
\item{curve.y}{y values}
}
\author{
John Ruscio <ruscio@tcnj.edu> and Shirley Wang <wangs7@tcnj.edu>
Maintainer: John Ruscio <ruscio@tcnj.edu>
}
