\name{CalculateProfileOutput}
\alias{CalculateProfileOutput}
\title{
Provides aggregated CCFIs and base rate estimates for CCFI profile
}
\description{
This function provides aggregated CCFIs and base rate estimates for
CCFI profile
}
\usage{
CalculateProfileOutput(CCFIs, parameters)
}
\arguments{
  \item{CCFIs}{
CCFI values across base rates and procedures
}
  \item{parameters}{
The data and program parameters
}
}
\details{
Called by higher-order functions; users do not need to call this 
function directly.
}
\value{
This function returns aggregated CCFI values.
}
\author{
John Ruscio <ruscio@tcnj.edu> and Shirley Wang <shirleywang@g.harvard.edu>
Maintainer: John Ruscio <ruscio@tcnj.edu>
}
