\name{CalculateSkew}
\alias{CalculateSkew}
\title{
Calculates skew
}
\description{
This function calculates the sample skewness of a distribution
}
\usage{
CalculateSkew(x)
}
\arguments{
  \item{x}{
The data vector
}
}
\details{
Called by higher-order functions; users do not need to call this 
function directly.
}
\value{
The sample skewness of x
}
\author{
John Ruscio <ruscio@tcnj.edu> and Shirley Wang <wangs7@tcnj.edu>
Maintainer: John Ruscio <ruscio@tcnj.edu>
}
