\name{CalculateValidity}
\alias{CalculateValidity}
\title{
Calculates Validity
}
\description{
This function calculates the standardized mean difference between two
groups (Cohen's D)
}
\usage{
CalculateValidity(x.1, x.2)
}
\arguments{
  \item{x.1}{
Data for the first group
}
  \item{x.2}{
Data for the second group
}
}
\details{
Called by higher-order functions; users do not need to call this 
function directly.
}
\value{
The standardized mean difference between groups
}
\author{
John Ruscio <ruscio@tcnj.edu> and Shirley Wang <shirleywang@g.harvard.edu>
Maintainer: John Ruscio <ruscio@tcnj.edu>
}
