\name{CheckClassification}
\alias{CheckClassification}
\title{
Checks classification
}
\description{
This function checks classification for problems, and terminates the 
program if necessary
}
\usage{
CheckClassification(group, n)
}
\arguments{
  \item{group}{
Classification of cases
}
  \item{n}{
Sample size
}
}
\details{
Called by higher-order functions; users do not need to call this 
function directly.
}
\value{
Nothing; text output if problem occurs
}
\author{
John Ruscio <ruscio@tcnj.edu> and Shirley Wang <wangs7@tcnj.edu>
Maintainer: John Ruscio <ruscio@tcnj.edu>
}
