\name{CheckParameters}
\alias{CheckParameters}
\title{
Checks parameters
}
\description{
This function checks the parameter specifications for problems, and 
adjusts these parameters as needed.
}
\usage{
CheckParameters(x, parameters)
}
\arguments{
  \item{x}{
The data matrix
}
  \item{parameters}{
The data and program parameters
}
}
\details{
Called by higher-order functions; users do not need to call this 
function directly.
}
\value{
Data parameters, adjusted as needed
}
\author{
John Ruscio <ruscio@tcnj.edu> and Shirley Wang <wangs7@tcnj.edu>
Maintainer: John Ruscio <ruscio@tcnj.edu>
}
