\name{DisplayPanels}
\alias{DisplayPanels}
\title{
Displays panels of graphs
}
\description{
This function provides panels of graphs for taxometric analysis
}
\usage{
DisplayPanels(x.results, x.dim.results, x.cat.results, parameters)
}
\arguments{
  \item{x.results}{
Empirical data results
}
  \item{x.dim.results}{
Dimensional comparison data results
}
  \item{x.cat.results}{
Categorical comparison data results
}
  \item{parameters}{
The data and program parameters
}
}
\details{
Called by higher-order functions; users do not need to call this 
function directly.
}
\value{
This function returns nothing, and provides graphical output only
}
\author{
John Ruscio <ruscio@tcnj.edu> and Shirley Wang <shirleywang@g.harvard.edu>
Maintainer: John Ruscio <ruscio@tcnj.edu>
}
