\name{DisplayProfiles}
\alias{DisplayProfiles}
\title{
Plots CCFI Profiles
}
\description{
This function plots CCFI profiles
}
\usage{
DisplayProfiles(CCFIs, parameters)
}
\arguments{
  \item{CCFIs}{
CCFI values across base rates and procedures
}
  \item{parameters}{
The data and program parameters
}
}
\details{
Called by higher-order functions; users do not need to call this 
function directly.
}
\value{
This function returns nothing, and provides graphical output only
}
\author{
John Ruscio <ruscio@tcnj.edu> and Shirley Wang <shirleywang@g.harvard.edu>
Maintainer: John Ruscio <ruscio@tcnj.edu>
}
