\name{DisplayTextOutput}
\alias{DisplayTextOutput}
\title{
Displays CCFI profile output
}
\description{
This function displays aggregated CCFIs and base rate estimates for
CCFI profile
}
\usage{
DisplayTextOutput(CCFIs, parameters)
}
\arguments{
  \item{CCFIs}{
CCFI values across base rates and procedures
}
  \item{parameters}{
The data and program parameters
}
}
\details{
Called by higher-order functions; users do not need to call this 
function directly.
}
\value{
This function returns nothing, and provides text output only
}
\author{
John Ruscio <ruscio@tcnj.edu> and Shirley Wang <wangs7@tcnj.edu>
Maintainer: John Ruscio <ruscio@tcnj.edu>
}
