\name{EstimateMAMBAC}
\alias{EstimateMAMBAC}
\title{
Estimates MAMBAC base rate
}
\description{
This function estimates the taxon base rate for a MAMBAC curve
}
\usage{
EstimateMAMBAC(curve)
}
\arguments{
  \item{curve}{
MAMBAC curve
}
}
\details{
Called by higher-order functions; users do not need to call this 
function directly.
}
\value{
Base rate estimate
}
\author{
John Ruscio <ruscio@tcnj.edu> and Shirley Wang <wangs7@tcnj.edu>
Maintainer: John Ruscio <ruscio@tcnj.edu>
}