\name{EstimateMAXEIG}
\alias{EstimateMAXEIG}
\title{
Estimates MAXEIG base rate
}
\description{
This function estimates the taxon base rate for a MAXEIG curve
}
\usage{
EstimateMAXEIG(curve)
}
\arguments{
  \item{curve}{
MAXEIG curve
}
}
\details{
Called by higher-order functions; users do not need to call this 
function directly.
}
\value{
Base rate estiamte
}
\author{
John Ruscio <ruscio@tcnj.edu> and Shirley Wang <shirleywang@g.harvard.edu>
Maintainer: John Ruscio <ruscio@tcnj.edu>
}
