\name{EstimateMAXSLOPE}
\alias{EstimateMAXSLOPE}
\title{
Estimates MAXSLOPE base rate
}
\description{
This function estimates the taxon base rate for a MAXSLOPE curve
}
\usage{
EstimateMAXSLOPE(curve.x, curve.y)
}
\arguments{
  \item{curve.x}{
X values of curve
}
  \item{curve.y}{
Y values of curve
}
}
\details{
Called by higher-order functions; users do not need to call this 
function directly.
}
\value{
Base rate estimate
}
\author{
John Ruscio <ruscio@tcnj.edu> and Shirley Wang <shirleywang@g.harvard.edu>
Maintainer: John Ruscio <ruscio@tcnj.edu>
}