\name{GetSpecifications}
\alias{GetSpecifications}
\title{
Provides analytic specifications
}
\description{
This function provides analytic specifications
}
\usage{
GetSpecifications(parameters)
}
\arguments{
  \item{parameters}{
The data and program parameters
}
}
\details{
Called by higher-order functions; users do not need to call this 
function directly.
}
\value{
This function returns nothing, and provides text output only
}
\author{
John Ruscio <ruscio@tcnj.edu> and Shirley Wang <shirleywang@g.harvard.edu>
Maintainer: John Ruscio <ruscio@tcnj.edu>
}
