\name{RemoveMissingData}
\alias{RemoveMissingData}
\title{
Removes missing data
}
\description{
This function performs listwise deletion of missing data
}
\usage{
RemoveMissingData(x)
}
\arguments{
  \item{x}{
The data matrix
}
}
\details{
Called by higher-order functions; users do not need to call this 
function directly.
}
\value{
Data after listwise deletion of missing data
}
\author{
John Ruscio <ruscio@tcnj.edu> and Shirley Wang <wangs7@tcnj.edu>
Maintainer: John Ruscio <ruscio@tcnj.edu>
}
