\name{RunMAMBAC}
\alias{RunMAMBAC}
\title{
Performs MAMBAC 
}
\description{
This function performs the MAMBAC analysis
}
\usage{
RunMAMBAC(x, parameters)
}
\arguments{
  \item{x}{
The data matrix
}
  \item{parameters}{
The data and program parameters
}
}
\details{
Called by higher-order functions; users do not need to call this 
function directly.
}
\value{
Panel of MAMBAC curves
}
\references{
Meehl, P.E., & Yonce, L.J. (1994). Taxometric analysis: I. Detecting
taxonomy with two quantitative indicators using means above and below a
sliding cut (MAMBAC procedure). Psychological Reports, 74(3, Pt 2), 1059-1274.
}
\author{
John Ruscio <ruscio@tcnj.edu> and Shirley Wang <shirleywang@g.harvard.edu>
Maintainer: John Ruscio <ruscio@tcnj.edu>
}