\name{RunMAXSLOPE}
\alias{RunMAXSLOPE}
\title{
Performs MAXSLOPE
}
\description{
This function performs the MAXSLOPE analysis
}
\usage{
RunMAXSLOPE(x)
}
\arguments{
  \item{x}{
The data matrix
}
}
\details{
Called by higher-order functions; users do not need to call this 
function directly.
}
\value{
Panel of MAXSLOPE curves
\item{curve.x }{X values of curve}
\item{curve.y }{Y values of curve}
}
\references{
Grove, W.M., & Meehl, P.E. (1993). Simple regression-based procedures for 
taxometric investigations. Psychological Reports, 73, 707-737.
}
\author{
John Ruscio <ruscio@tcnj.edu> and Shirley Wang <wangs7@tcnj.edu>
Maintainer: John Ruscio <ruscio@tcnj.edu>
}