\name{SummarizeDist}
\alias{SummarizeDist}
\title{
Summarizes distribution 
}
\description{
This function calculates the sample mean, standard deviation, skewness, 
and kurtosis
}
\usage{
SummarizeDist(x)
}
\arguments{
  \item{x}{
The data vector
}
}
\details{
Called by higher-order functions; users do not need to call this 
function directly.
}
\value{
The sample mean, standard deviation, skewness, and kurtosis of x.
}
\author{
John Ruscio <ruscio@tcnj.edu> and Shirley Wang <wangs7@tcnj.edu>
Maintainer: John Ruscio <ruscio@tcnj.edu>
}
