\name{DisplayBaseRates}
\alias{DisplayBaseRates}
\title{
Displays base rates
}
\description{
This function calculates and reports base rate estimates for taxometric
analysis.
}
\usage{
DisplayBaseRates(x.results, parameters)
}
\arguments{
  \item{x.results}{
Empirical data results
}
  \item{parameters}{
The data and program parameters
}
}
\details{
Called by higher-order functions; users do not need to call this 
function directly.
}
\value{
This function returns nothing, and provides text output only
}
\author{
John Ruscio <ruscio@tcnj.edu> and Shirley Wang <wangs7@tcnj.edu>
Maintainer: John Ruscio <ruscio@tcnj.edu>
}
