\name{DisplayCCFIs}
\alias{DisplayCCFIs}
\title{
Calculates and displays CCFIs 
}
\description{
This function calculates and reports CCFI results for taxometric analysis
}
\usage{
DisplayCCFIs(x.results, x.dim.results, x.cat.results, parameters)
}
\arguments{
  \item{x.results}{
Empirical data results
}
  \item{x.dim.results}{
Dimensional comparison data results
}
  \item{x.cat.results}{
Categorical comparison data results
}
  \item{parameters}{
The data and program parameters
}
}
\details{
Called by higher-order functions; users do not need to call this 
function directly.
}
\value{
This function returns nothing, and provides text output only
}
\author{
John Ruscio <ruscio@tcnj.edu> and Shirley Wang <wangs7@tcnj.edu>
Maintainer: John Ruscio <ruscio@tcnj.edu>
}
